Extract course-related helpers to course-helpers.ts
