+++ apps/antalmanac/src/components/RightPane/SectionTable/SectionTableBody.tsx
@@ -26,7 +26,8 @@ import { ColorAndDelete, ScheduleAddCell } from './SectionTableButtons';
 import restrictionsMapping from './static/restrictionsMapping.json';
 import GradesPopup from './GradesPopup';
 import analyticsEnum, { logAnalytics } from '$lib/analytics';
-import { clickToCopy, CourseDetails, isDarkMode } from '$lib/helpers';
+import { clickToCopy, isDarkMode } from '$lib/helpers';
+import { CourseDetails } from '$lib/course_data.types';
 import Grades from '$lib/grades';
 import AppStore from '$stores/AppStore';
 import { useTabStore } from '$stores/TabStore';
