+++ apps/antalmanac/src/components/RightPane/SectionTable/SectionTableButtons.tsx
@@ -9,7 +9,7 @@ import { MOBILE_BREAKPOINT } from '../../../globals';
 import { addCourse, deleteCourse, openSnackbar } from '$actions/AppStoreActions';
 import ColorPicker from '$components/ColorPicker';
 import analyticsEnum, { logAnalytics } from '$lib/analytics';
-import { CourseDetails } from '$lib/helpers';
+import { CourseDetails } from '$lib/course_data.types';
 import AppStore from '$stores/AppStore';
 
 // Reset these params in url becasue when copy a specific class's link, it only copy its course code
