+++ apps/antalmanac/src/lib/course_data.types.ts
@@ -1,0 +1,15 @@
+import { WebsocSection } from "peterportal-api-next-types";
+
+
+export interface CourseDetails {
+    deptCode: string;
+    courseNumber: string;
+    courseTitle: string;
+    courseComment: string;
+    prerequisiteLink: string;
+}
+
+export interface CourseInfo {
+    courseDetails: CourseDetails;
+    section: WebsocSection;
+}
